feat: implement CSS selector extraction with regex fallback
- Add cheerio for structured HTML parsing
- Prioritize CSS selectors in research agent and discovery pipeline
- Refactor research agent and discovery modules to comply with 500-line file limit
- Add unit tests for HTML extraction utilities
- Resolve dependency vulnerabilities and type safety issues
