+++ package.json
@@ -34,8 +34,7 @@
   "dependencies": {
     "cheerio": "^1.2.0",
     "protobufjs": "8.3.0",
-    "zod": "^3.22.4",
-    "sanitize-html": "^2.17.4"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "^0.16.6",
@@ -44,6 +43,7 @@
     "@vitest/coverage-v8": "^4.1.6",
     "artillery": "^2.0.31",
     "miniflare": "^4.20260515.0",
+    "minimatch": "^10.2.5",
     "prettier": "^3.8.3",
     "typescript": "^6.0.3",
     "vitest": "^4.1.5",
@@ -54,7 +54,8 @@
     "fast-xml-builder": "1.2.0",
     "protobufjs": "8.3.0",
     "uuid": "14.0.0",
-    "@playwright/test": "^1.60.0"
+    "@playwright/test": "^1.60.0",
+    "ws": "8.20.1"
   },
   "engines": {
     "node": ">=22.0.0"
