+++ worker/lib/research-agent/api-fetchers.ts
@@ -34,8 +34,8 @@ export async function fetchProductHuntDeals(
   }
 
   const query = `
-    query ($limit: Int!, $searchQuery: String!) {
-      posts(first: $limit, order: RANKING, search: {query: $searchQuery}) {
+    query {
+      posts(first: ${limit}, order: RANKING, search: {query: "${searchQuery.replace(/"/g, '\\"')}"}) {
         edges {
           node {
             id
@@ -69,13 +69,7 @@ export async function fetchProductHuntDeals(
         "Content-Type": "application/json",
         Accept: "application/json",
       },
-      body: JSON.stringify({
-        query,
-        variables: {
-          limit,
-          searchQuery,
-        },
-      }),
+      body: JSON.stringify({ query }),
       signal: AbortSignal.timeout(10000),
     });
